pyproject.toml
[project]
name = "gcloud-logs-anomaly-detection"
version = "0.1.0"
description = "Machine Learning tool to find anomalies in Google Cloud Logging events"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "23skdu", email = "23skdu@users.noreply.github.com"}
]
dependencies = [
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "google-cloud-logging>=3.12.1",
    "seaborn>=0.13.2",
    "matplotlib>=3.10.0",
    "langchain>=0.3.17",
    "langchain-google-genai>=1.0.9",
    "langchain-ollama>=0.2.1",
    "lorem-text>=2.1",
    "google-api-core>=2.24.2",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.9.7",
    "mypy>=1.15.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100
lint.select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
lint.ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_calls = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
